pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "intraday-cad"
version = "0.1.0"
description = "OR / Gap / MM probability dashboard feature-engineering and analytics pipeline."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Proprietary" }
authors = [
  { name = "Mehmet Bilgili" }
]
keywords = ["trading", "intraday", "analytics", "streamlit", "pandas"]
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "pyyaml>=6.0",
  "openpyxl>=3.1",
  "pyarrow>=14.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.6",
]
app = [
  "streamlit>=1.36",
  "plotly>=5.20",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["data_io*", "features*", "analytics*", "app*", "reports*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I"]
